requirements.txt
torch>=2.0.0
transformers==4.40.2
diffusers==0.27.2
huggingface-hub==0.20.2
Pillow
numpy
nltk
flax
jax[cpu]
